+++ MasjidOnline.Entity/CaptchaAnswer.cs
@@ -4,9 +4,9 @@ namespace MasjidOnline.Entity;
 
 public class CaptchaAnswer
 {
-    public required int Id { get; set; }
+    public required long Id { get; set; }
 
-    public required int CaptchaQuestionId { get; set; }
+    public required long CaptchaQuestionId { get; set; }
 
     public required float Degree { get; set; }
