+++ MasjidOnline.Entity/Donation.cs
@@ -4,11 +4,11 @@ namespace MasjidOnline.Entity;
 
 public class Donation
 {
-    public required int Id { get; set; }
+    public required long Id { get; set; }
 
     public required DateTime DateTime { get; set; }
 
-    public int? UserId { get; set; }
+    public long? UserId { get; set; }
 
     public string? AnonymousName { get; set; }
 }
